requirements.txt
mysql-connector-python
gym